fixed bugs in functions
